feat(validate): 로또 번호 유효성검증 로직 교체
